% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mems.r
\name{bound_half_segments}
\alias{bound_half_segments}
\title{Function to find which segments on boundary}
\usage{
bound_half_segments(mesh)
}
\arguments{
\item{mesh}{A \code{fmesher::fm_mesh_2d} object.}
}
\description{
Function to find which segments on boundary
}
