% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mems.r
\name{cens}
\alias{cens}
\title{Function to get mesh triangle center points of a \code{mesh} object}
\usage{
cens(mesh)
}
\arguments{
\item{mesh}{A \code{fmesher::fm_mesh_2d} object.}
}
\description{
Function to get mesh triangle center points of a \code{mesh} object
}
\details{
Used for plotting label locations
}
