% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{gg_mems}
\alias{gg_mems}
\title{Plots the calculated metrics from a \code{mems} output
(i.e., each triangle metric).}
\usage{
gg_mems(x, metric = "m1")
}
\arguments{
\item{x}{output of \code{mems::mems}}

\item{metric}{character specifying which metric of \code{x} to plot one of "metric_1",
"metric_2", "metric_3", "metric_4", "metric_5", "metric_6"  (default "metric_1").}
}
\value{
A  \code{"ggplot"} object.
}
\description{
Plots the calculated metrics from a \code{mems} output
(i.e., each triangle metric).
}
