% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{gg_tri}
\alias{gg_tri}
\title{Plot triangle with all the bells
internal function only}
\usage{
gg_tri(
  A,
  B,
  C,
  node_labels = c("A", "B", "C"),
  edge_labels = c("a", "b", "c"),
  r = c(0.2, 0.2, 0.2),
  ang_labels = c(expression(theta[a]), expression(theta[b]), expression(theta[c])),
  circum = FALSE,
  incircle = FALSE
)
}
\description{
Plot triangle with all the bells
internal function only
}
