% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mems.r
\name{half_segments}
\alias{half_segments}
\title{Function to find the half way point of each
mesh edge and turn into an \code{sf} \code{LINESTRING}
object}
\usage{
half_segments(mesh)
}
\arguments{
\item{mesh}{A \code{fmesher::fm_mesh_2d} object.}
}
\description{
Function to find the half way point of each
mesh edge and turn into an \code{sf} \code{LINESTRING}
object
}
