% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mems.r
\name{mems}
\alias{mems}
\title{Calculate a number of different geometric attributes of a Delaunay triangulation}
\usage{
mems(mesh, plot = FALSE)
}
\arguments{
\item{mesh}{A \code{fmesher::fm_mesh_2d} object.}

\item{plot}{If \code{TRUE} then triangle quality metrics are plotted.}
}
\value{
An object of class \code{sf} with the following data for each triangle in the
triangulation
\itemize{
\item \code{V1}, \code{V2}, and \code{V3} corresponding vertices
of \code{mesh} matches \code{mesh$graph$tv};
\item \code{ID}, numeric triangle id;
\item \code{angleA}, \code{angleB}, and \code{angleC}, the
interior angles;
\item circumcircle radius, circumradius, \code{circumcircle_R} (\eqn{R});
\item incircle radius \code{incircle_r} (\eqn{r});
\item centroid locations of the circumcircle, circumcenter, (\code{c_Ox, c_Oy});
\item centroid locations of the incircle, incenter, (\code{i_Ox, i_Oy});
\item the radius-edge ratio \code{radius_edge} \eqn{\frac{R}{l_{min}}},
where \eqn{l_{min}} is the minimum edge length;
\item the radius ratio \code{radius_ratio} \eqn{\frac{r}{R}};
\item \code{area}, area (\eqn{A});
\item \code{metric_c1}: product of each triangle's nodes rowsums of the C1 matrix
\item \code{metric__g1}: product of each triangle's nodes rowsums of the G1 matrix
\item \code{metric_b1}: summation of each triangle's nodes rowsums of the B1 matrix
\item \code{metric_1}: \eqn{\frac{4\sqrt{3}|A|}{\Sigma_{i = 1}^3 L_i^2}},
where \eqn{L_i} is the length of edge \eqn{i}.
\item \code{metric_2}: \eqn{6 \sqrt{\frac{A}{\sqrt{3}}}}.
\item \code{metric_3}:\eqn{\frac{\text{min}(L_i)}{\text{max}(L_i)}}.
\item \code{metric_4}:  \eqn{3 \frac{\text{min}(\phi_i)}{\pi}}, where
\eqn{\phi_i} is the \eqn{i^{th}} internal angle.
\item \code{metric_5}: \eqn{\frac{4\sqrt{3} A}{\text{max}(L_i)} * \Sigma_{i = 1}^3 L_i^2}.
\item \code{metric_6}: \eqn{\frac{1}{4 q_b} + q_w \frac{\sqrt{3}}{16}}, where
\eqn{q_b = \frac{\Sigma_{i = 1}^3 L_i^2}{4 \sqrt{3} A}} and
\eqn{q_m = (\frac{1}{3A})(\Sigma_{i = 1}^3 L_i)^2}.
}
}
\description{
Calculates a number of geometric attributes for a given
Delaunay triangulation based on the circumscribed and inscribed circle of each triangle.
}
\details{
A triangle's circumcircle (circumscribed circle) is the unique circle that passes
through each of its three vertices. A triangle's incircle (inscribed circle) is the
largest circle that can be contained within it (i.e., touches it's three edges).
}
\examples{
data(example_mesh, package = "mems")
metrics <- mems(example_mesh)
}
