% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mems.r
\name{mesh_2_sf}
\alias{mesh_2_sf}
\title{Transform a \code{fmesher::fm_mesh_2d} into a \code{sf} object}
\source{
Modified from \code{sp} based function suggested by Finn in the
R-inla discussion Google Group
\url{https://groups.google.com/g/r-inla-discussion-group/c/z1n1exlZrKM}.
}
\usage{
mesh_2_sf(mesh)
}
\arguments{
\item{mesh}{A \code{fmesher::fm_mesh_2d} object.}
}
\value{
A simple features, \code{sf}, object.
}
\description{
Transform a \code{fmesher::fm_mesh_2d} into a \code{sf} object
}
\examples{
data(example_mesh)
sf <- mesh_2_sf(example_mesh)
}
\seealso{
\code{\link{mems}}
}
