% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mems.r
\name{mesh_ang}
\alias{mesh_ang}
\title{Calculate all interior Delaunay triangulation triangle angles}
\usage{
mesh_ang(mesh)
}
\arguments{
\item{mesh}{A \code{fmesher::fm_mesh_2d} object.}
}
\description{
Calculate all interior Delaunay triangulation triangle angles
}
\examples{
data(example_mesh)
mesh_ang(example_mesh)
}
