% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mems.r
\name{metrics}
\alias{metrics}
\title{A range of triangle quality metrics and measures}
\usage{
metrics(A, B, C)
}
\arguments{
\item{A}{(x,y) coords of vertex A}

\item{B}{(x,y) coords of vertex B}

\item{C}{(x,y) coords of vertex C}
}
\description{
A range of triangle quality metrics and measures
}
