% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mems.r
\name{mid}
\alias{mid}
\title{Find mid point on straight line between two points}
\usage{
mid(A, B)
}
\arguments{
\item{A}{A numeric vector of length 2 specifying (first) location.}

\item{B}{A numeric vector of length 2 specifying (second) location.}
}
\description{
Find mid point on straight line between two points
}
\examples{
mid(c(1, 1), c(2, 1))
}
