% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_blocks.r
\name{q_blocks}
\alias{q_blocks}
\title{Calculates the precision matrix blocks contributions from each triangle/node
in the triangulation}
\source{
\url{https://github.com/inlabru-org/fmesher/blob/d1dcbba1a42144a22b8e8056cb7236381e0b67ef/src/mesh.cc#L2278}
}
\usage{
q_blocks(mesh)
}
\arguments{
\item{mesh}{object of \code{fmesher::fm_mesh_2d} (i.e., 2D Delaunay triangulation, 'mesh').}
}
\description{
Calculates the precision matrix blocks contributions from each triangle/node
in the triangulation
}
\details{
Simpler version of the \code{C++} versions implemented in
\code{fmesher::fm_fem}. Is only for demonstration.
}
