% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mems.r
\name{segments}
\alias{segments}
\title{Extract a dataframe of mesh triangle segments (start and end locations)}
\usage{
segments(mesh)
}
\arguments{
\item{mesh}{A \code{fmesher::fm_mesh_2d} object.}
}
\description{
Extract a dataframe of mesh triangle segments (start and end locations)
}
\examples{
data(example_mesh)
segments(example_mesh)
}
