% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{gg_mems}
\alias{gg_mems}
\title{Plots the calculated metrics from a \code{mems} output
(i.e., each triangle metric).}
\usage{
gg_mems(x)
}
\value{
A  \code{"patchwork"}, \code{"gg"}, \code{"ggplot"} object.
}
\description{
Plots the calculated metrics from a \code{mems} output
(i.e., each triangle metric).
}
