% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{gg_seg_influence}
\alias{gg_seg_influence}
\title{Plot segment influence
only useful for small number of nodes}
\usage{
gg_seg_influence(mesh)
}
\arguments{
\item{mesh}{A \code{fmesher::fm_mesh_2d} object.}
}
\description{
Plot segment influence
only useful for small number of nodes
}
