% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mems.r
\name{metrics}
\alias{metrics}
\title{A range of triangle quality metrics}
\usage{
metrics(a, b, c)
}
\arguments{
\item{a}{A numeric triangle edgelength (opposite angle A).}

\item{b}{A numeric triangle edgelength (opposite angle B).}

\item{c}{A numeric triangle edgelength (opposite angle C).}
}
\description{
A range of triangle quality metrics
}
